[RemoteInspection] Parse a Builtin.Int's bit width into an unsigned.
intTypeBitSize parsed the width out of a 'Bi' N '_' mangling into a uint8_t, so anything at or above Bi256_ overflowed, getAsInteger reported failure, and the caller concluded the type was not an integer at all and read its extra inhabitants as if it were a heap pointer. Nothing wider than Builtin.Int128 is reachable today, and Builtin.Int256 gets no builtin descriptor to lower from, so this is a latent case rather than a live bug.

+++ stdlib/public/RemoteInspection/TypeLowering.cpp
@@ -277,7 +277,7 @@ static unsigned intTypeBitSize(std::string name) {
   llvm::StringRef nameRef(name);
   if (nameRef.starts_with("Bi") && nameRef.ends_with("_")) {
     llvm::StringRef naturalRef = nameRef.drop_front(2).drop_back();
-    uint8_t natural;
+    unsigned natural;
     if (naturalRef.getAsInteger(10, natural)) {
       return 0;
     }
